Reimplement User model using Django's built-in User
The original implementation reinvented the wheel unnecessarily. The new
implementation takes advantage of Django's built in 'User' fields.

A 'Profile' is automatically created for each new User. This profile
stores the custom additional 'Favourites' field.
Access is performed with 'user.profile.favourites'.

The admin.py file is also updated so that the admin panel displays user
favourites.

The population script is also updated to handle the new User model
implementation.