Revise the controversial algorithm
Sandwiches with greater than a certain number of likes and dislikes are
considered for controversy.
A percentage difference between likes and dislikes is then calculated
and the sandwiches with less than a certain percentage are considered to
be controversial.
These remaining sandwiches are then sorted in order of closest matching
likes and dislikes.